# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoRigo AI Business Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orchestrate:
    post:
      operationId: orchestrate
      tags: [orchestrate]
      summary: Voice command orchestration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrchestrateRequest"
      responses:
        "200":
          description: Orchestration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrchestrateResponse"

  # ───── Sessions ─────
...
  /notifications/push-token:
    post:
      operationId: registerPushToken
      tags: [notifications]
      summary: Register Expo push token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenRequest"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removePushToken
      tags: [notifications]
      summary: Remove push token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required: [token]
      responses:
        "200":
          description: Token removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      summary: Get notification preferences per business
      parameters:
        - name: businessId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    put:
      operationId: updateNotificationPreferences
      tags: [notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesRequest"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"

  # ───── Referrals ─────
...
  /knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      summary: List knowledge documents for a business
      parameters:
        - name: businessId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Knowledge documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeDocument"
    post:
      operationId: uploadKnowledge
      tags: [knowledge]
      summary: Upload a document to the knowledge base
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                businessId:
                  type: string
                title:
                  type: string
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeDocument"
  /knowledge/{id}:
    delete:
      operationId: deleteKnowledge
      tags: [knowledge]
      summary: Delete a knowledge document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ───── Account ─────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required: [success]

    # ── Business ──
...
    OrchestrateRequest:
      type: object
      properties:
        message:
          type: string
        businessId:
          type: string
        sessionMode:
          type: ["string", "null"]
          enum: [deep_research, strategy_swot, brainstorm, business_plan, "null"]
        provider:
          type: ["string", "null"]
          enum: [openai, anthropic, "null"]
      required: [message, businessId]
    OrchestrateResponse:
      type: object
      properties:
        response:
          type: string
        sessionMode:
          type: ["string", "null"]
        sessionStep:
          type: ["integer", "null"]
        sessionTotalSteps:
          type: ["integer", "null"]
        toolsUsed:
          type: array
          items:
            type: string
        usage:
          type: ["integer", "null"]
      required: [response, toolsUsed]

    # ── Sessions ──
...
    PushTokenRequest:
      type: object
      properties:
        token:
          type: string
        platform:
          type: string
          enum: [ios, android]
      required: [token, platform]
    NotificationPreferences:
      type: object
      properties:
        businessId:
          type: string
        agentActivity:
          type: boolean
        communications:
          type: boolean
        billingAlerts:
          type: boolean
      required: [businessId, agentActivity, communications, billingAlerts]
    UpdateNotificationPreferencesRequest:
      type: object
      properties:
        businessId:
          type: string
        agentActivity:
          type: ["boolean", "null"]
        communications:
          type: ["boolean", "null"]
        billingAlerts:
          type: ["boolean", "null"]
      required: [businessId]

    # ── Referrals ──
...
    KnowledgeDocument:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        businessId:
          type: string
        status:
          type: string
          enum: [processing, ready, failed]
        chunkCount:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, title, businessId, status, createdAt]

    # ── Auth ──
...